Move EnsDeployer to ops
