+++ universal-login-ops/package.json
@@ -16,13 +16,14 @@
     "clean": "rm -fr dist"
   },
   "dependencies": {
-    "ethereum-waffle": "2.0.11",
-    "ganache-core": "^2.2.1",
-    "knex": "^0.16.3",
     "@universal-login/commons": "0.2.0",
     "@universal-login/contracts": "^0.2.0",
     "@universal-login/relayer": "^0.2.0",
     "@universal-login/sdk": "^0.2.0",
+    "ens-builder": "^2.0.1",
+    "ethereum-waffle": "2.0.11",
+    "ganache-core": "^2.2.1",
+    "knex": "^0.16.3",
     "yargs": "^12.0.5"
   },
   "resolutions": {
